Menu is about 80-90% done
rudimentary numeric system is almost finished, as of now the menu should
work defaulting to port 8000 addr 127.0.0.1
all the other argument options such as ship color and host type should
work now, i,e if host type is set to peer (client), the menu will not
run the server program. Also there is a global option to mute and unmute
the bgm, which pauses in position when it is clicked.